+++ src/DeterministicFiniteStateMachineInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Dhii\Machine;
+
+/**
+ * Something that has a current state, a finite set of possible states and a finite set of
+ * transitions that can be applied from each state.
+ *
+ * @since [*next-version*]
+ */
+interface DeterministicFiniteStateMachineInterface extends
+    FiniteStateMachineInterface,
+    TransitionsAwareInterface,
+    PossibleTransitionsAwareInterface
+{
+}