+++ src/TransitionCapableInterface.php
@@ -2,8 +2,10 @@
 
 namespace Dhii\Machine;
 
+use Dhii\Data\ValueAwareInterface;
+
 /**
- * Something that can transition from one state to another.
+ * Something that can have a transition applied to change its state.
  *
  * @since [*next-version*]
  */
@@ -14,9 +16,9 @@ interface TransitionCapableInterface
      *
      * @since [*next-version*]
      *
-     * @param TransitionInterface $transition The transition to apply.
+     * @param ValueAwareInterface $transition The transition to apply.
      *
-     * @return bool True if the transition was successful, false if not.
+     * @throws TransitionExceptionInterface If the transition failed.
      */
-    public function transition(TransitionInterface $transition);
+    public function transition(ValueAwareInterface $transition);
 }