+++ test/functional/StateTransitionCapableInterfaceTest.php
@@ -2,34 +2,34 @@
 
 namespace Dhii\Machine\FuncTest;
 
-use \Dhii\Machine\StateInterface;
+use \Dhii\Machine\StateTransitionCapableInterface;
 use \Xpmock\TestCase;
 
 /**
- * Tests {@see Dhii\Machine\StateInterface}.
+ * Tests {@see Dhii\Machine\StateTransitionCapableInterface}.
  *
  * @since [*next-version*]
  */
-class StateInterfaceTest extends TestCase
+class StateTransitionCapableInterfaceTest extends TestCase
 {
     /**
      * The name of the test subject.
      *
      * @since [*next-version*]
      */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\\Machine\\StateInterface';
+    const TEST_SUBJECT_CLASSNAME = 'Dhii\\Machine\\StateTransitionCapableInterface';
 
     /**
      * Creates a new instance of the test subject.
      *
      * @since [*next-version*]
      *
-     * @return StateInterface
+     * @return StateTransitionCapableInterface
      */
     public function createInstance()
     {
         $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
-            ->getValue()
+            ->transitionTo()
             ->new();
 
         return $mock;
@@ -46,9 +46,6 @@ class StateInterfaceTest extends TestCase
 
         $this->assertInstanceOf(
             static::TEST_SUBJECT_CLASSNAME, $subject, 'Subject is not a valid instance.'
-        );
-        $this->assertInstanceOf(
-            'Dhii\\Data\\ValueAwareInterface', $subject, 'Subject is not a valid instance.'
         );
     }
 }