+++ test/functional/StateTransitionableInterfaceTest.php
@@ -2,34 +2,35 @@
 
 namespace Dhii\Machine\FuncTest;
 
-use \Dhii\Machine\StateMachineInterface;
-use \Xpmock\TestCase;
+use Dhii\Machine\StateTransitionableInterface;
+use Xpmock\TestCase;
 
 /**
- * Tests {@see Dhii\Machine\StateMachineInterface}.
+ * Tests {@see Dhii\Machine\StateTransitionableInterface}.
  *
  * @since [*next-version*]
  */
-class StateMachineInterfaceTest extends TestCase
+class StateTransitionableInterfaceTest extends TestCase
 {
     /**
      * The name of the test subject.
      *
      * @since [*next-version*]
      */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\\Machine\\StateMachineInterface';
+    const TEST_SUBJECT_CLASSNAME = 'Dhii\\Machine\\StateTransitionableInterface';
 
     /**
      * Creates a new instance of the test subject.
      *
      * @since [*next-version*]
      *
-     * @return StateMachineInterface
+     * @return StateTransitionableInterface
      */
     public function createInstance()
     {
         $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
-            ->getCurrentState()
+            ->canTransitionTo()
+            ->transitionTo()
             ->new();
 
         return $mock;
@@ -48,7 +49,12 @@ class StateMachineInterfaceTest extends TestCase
             static::TEST_SUBJECT_CLASSNAME, $subject, 'Subject is not a valid instance.'
         );
         $this->assertInstanceOf(
-            'Dhii\\Machine\\CurrentStateAwareInterface', $subject, 'Subject is not a valid StateAwareInterface instance.'
+            'Dhii\\Machine\\StateTransitionCheckCapableInterface',
+            $subject, 'Subject is not a valid StateTransitionCheckCapableInterface instance.'
+        );
+        $this->assertInstanceOf(
+            'Dhii\\Machine\\StateTransitionCapableInterface',
+            $subject, 'Subject is not a valid StateTransitionCapableInterface instance.'
         );
     }
 }