+++ CHANGELOG.md
@@ -6,3 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 Initial release.
+
+### Added
+- Interop interfaces;
+- Interface tests.